Make mobilenet handle image resizing and normalization
Previously, this code used imgToTensor() to turn a resource into a tensor with values [-1, 1]. However, the code in tensorflow-models/mobilenet will itself normalize the input (expected to be [0, 255]) to [0, 1]. Applied twice, the resulting vector yielded the incorrect results seen in the previous commit.

This sidesteps the conversion to a tensor altogether to let tensorflow-models/mobilenet take care of it instead. (I haven't tested this with anything except MobileNet at this point.)

See also:
https://github.com/tensorflow/tfjs-models/blob/14acf0c3c675713186548355d91b5aa95730136a/mobilenet/src/index.ts#L218